Move existing canvas's to CanvasArenas folder
