+++ app/client/src/pages/common/CanvasArenas/CanvasMultiPointerArena.tsx
